ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow>=2.17.0
numpy>=1.26.0
Pillow>=10.2.0
gunicorn==21.2.0
